Refactoring of search and download functions
